login no business, compareHash não funciona
