+++ Semana 19 - Arquitetura de Software/Projeto - LaBook/src/business/user/loginBusiness.ts
@@ -1,41 +1,40 @@
 import { UserDatabase } from "../../data/user/UserDatabase";
 import { user } from "../../model/types";
+import { generateToken } from "../../services/authenticator";
+import { compareHash } from "../../services/hashManager";
 import { CustomError } from "../error/CustomError";
 
 const loginBusiness = async (user: user) => {
   try {
-    // const database = new UserDatabase();
-    console.log(user)
-
-    if(!user.email || !user.password){      
-      throw new CustomError(400, "The fields 'name' and 'password' are required.");
+    if (!user.email || !user.password) {
+      throw new CustomError(
+        400,
+        "The fields 'name' and 'password' are required."
+      );
     }
 
-    // const [result] = await database.selectUser(user.email);
-    // if(result.length === 0){
-    //   console.log({ a: "ta vaziooo"})
-    //   throw new CustomError(404, "No results");
-    // }
+    const database = new UserDatabase();
+    const [result] = await database.selectUser(user.email);
+    if (result.length === 0) {
+      throw new CustomError(404, "No results");
+    }
 
+    // const passwordIsCorrect: boolean = compareHash(
+    //   user.password,
+    //   result[0].password
+    // );
 
-    // const user: user = {
-    //   id: queryResult[0].id,
-    //   name: queryResult[0].name,
-    //   email: queryResult[0].email,
-    //   password: queryResult[0].password,
-    // };
-
-    // const passwordIsCorrect: boolean = await compare(password, user.password);
+    // console.log(passwordIsCorrect);
 
     // if (!passwordIsCorrect) {
-    //   res.statusCode = 401;
-    //   message = "Invalid credentials";
-    //   throw new Error(message);
+    //   throw new CustomError(401, "Password incorrect");
     // }
 
-    // const token: string = generateToken({
-    //   id: user.id,
-    // });
+    const token: string = generateToken({
+      id: result[0].id,
+    });
+
+    return token
   } catch (error) {
     throw new CustomError(error.statusCode, error.message);
   }
